Firebase calls (#11)
* Adding interface for uploader to firebase to be implemented

* Fixed missing fields in Caption, added User class, and fixed parameters of Uploader

* Fixed compile errors and constructor ordering for Caption

* Added photo to addUser which will be their profile picture

* Implemented the addGame method in FirebaseUploadMethods. Assumes that the
Game class already has a key (hence not using push()). Also made a
progress bar. Changed the FAB to call the method so it can be
demonstrated, and once the actual upload screen is implemented the FAB
will be changed anyways so I figured that was the best spot for it.

* Made changes requested by Austin in code review. Also implemented the
getNewCaptionKey class, and altered the Uploader interface a little by
adding a gameId parameter to getNewCaptionKey().

* fixed a bug in addGameToUserTable() whicj caused excessive resource usage