Recycler cards (#30)
* Adding all the stuff for actual user input and captioning. Now when a user
selects a card from the popup menu, it populates the text fields, and
allows the user to input something to the field. Still not sure how the
"submit" button should be, that will be next commit

* Doing some UI cleanup. Added a submit button that can be pressed instead
of enter on the keyboard.

* Forcing keyboard to show up/go away once a user selects a card/submits
their caption.

* Small changes to make code more clear

* Renaming CaptionActivity to GameActivity

* Changing from the menu popping up from the FAB, to being a recycler view
you can scroll sideways through. I think this way looks better, but will
need further UI polish, namely having the user have the ability to swipe
up from the bottom to bring up the cards.

* Removed the FAB and menu, instead having a recycler view with cards. Also
incorporated feedback from code review, and fixed a lot of sloppy code.
Wrote some tests too!

* Adding tests. Even though I said I added them in the last one, I actually
added them here.

* Added todo for verifying login status

* Adding FAB that can pop-up or remove the possible card list

* Changes to support JaCoCo testing. Couldn't figure out how to separate it
from my local branch, not worth it at this point

* Adding height to the caption cards so they have some nifty shadow.

* Removing leftover test suite. Changing method name.

* Removed some unnecessary todos and styles. Also made it so that you can
click on any part of the card to choose it, instead of only the text.