Adding deep link capability, and invites (#44)
* Initial work to set up facebook invites. Adding invite button and facebook-created dialog to find facebook friends.

* Now Facebook invite button is not visible if you are logged in on Google+

* Adding support for Firebase Dynamic Links. For now, only supports deep
linking to a game.
To deep link to a specific game, the url should be
https://snaptiongame.com/games/<GAME_ID> and the link can be generated by
calling FirebaseDeepLinker.getDeepLink()

* Removing stuff I didn't mean to pull in

* Adding button to send an invite message that includes a dynamic link to
our home screen if the user is on android, otherwise they will go to the
snaption website

* Removed some unnecessary printlns and extra stuff from trying things.
Added test, full FirebaseDeepLinker coverage aside from RetrofitError and
IOException lines

* Fixed an issue where the app wouldn't show up in the app drawer

* Making changes based on Code Review. Mainly, created new class
"DeepLinkGetter" which checks if a deep link is in the Main Activity, and
if it is, starts an intent from it. Also renamed FirebaseDeepLinker to
FirebaseDeepLinkCreator to be more clear. Also added a check for incorrect
game IDs.

* Further extracted the deep link check, so that DeepLinkGetter will now
also start the intent if need be, rather than passing it back to the
activity

* Added comment to DeepLinkGetter